ENH: A class for primal/dual structure and adaptor
itkQuadEdgeMeshWithDual is a prototype for a class which would handle primal and
dual in the same structure. It only handle the higher level interface (ITK
level) and does not handle sharing QuadEdges between the primal and the dual as
it should. The dual cells and points actually do not create or link to any QE.
This is sufficent for display.

An adaptor as been written so that the datastructure can be passed to a writer
through the adaptor and work out of the box.

TODO:
- infrastructure to handle sharing the QEs
  -- QE container? pack them by 4, keep the main offset, use modulo (see paper)
  -- QEPoint is fine
  -- QECell needs to be rewritten using the container (constructor)
  -- QEPolygon should be fine
- handle the prinal / dual QE types symetry
  -- check types, especially the adaptor
  -- can we avoid duplicating methods? macros (arggg)? one more template?
- new OriginRefType as a pair. What is the impact?
- new FaceRefType as a pair. what is the impact?
  -- the ITK layer should be fine
  -- the QE layer should be fine
  -- the QEGeom layer and itkQuadEdgeMesh should be highly impacted as I expect
     a lot of assumption there.

Change-Id: I274d1a5ef8267828fd05efdc0abc2f776972e025