ENH: put duplicated code in a function. Finish extracting code.
Code was simplified. Small functions were extracted. The main
is now only the layout of the algorithm. Next step is to
create a class from those functions.

Note on the method: extracting functions might be slightly slower
than writing a class directly. However, it allows for incremental
changes, and verification during the process. The scope of each
variable is assessed, and the functions make perfect methods
prototypes. There is extra work in the sense that most of the
types need to be declared again in each function, and the mesh
needs to be passed as arg even though it will be referenced either
by the this pointer or by GetOutput, but it is more than balanced
by the advantages listed above.

Change-Id: I53d6841c69eba9ee9903f29938c5877413977341